Extract webpack plugins to webpack/plugins
- Moving this config to its own folder allows us to break down this big file into
  smaller maintainable chunks we can use later on.
